add files classe
